requirements.txt
python-dateutil~=2.8.2
numpy~=1.25.0
boto3~=1.26.161
cron-converter~=1.0.2
pytest-html~=3.2.0
pytest~=7.4.0